requirements.txt
opencv-python>=4.8.0
pillow>=10.0.0
pyautogui>=0.9.54
pytesseract>=0.3.10
pyyaml>=6.0
numpy>=2.3.5
colorama>=0.4.6
pyinstaller>=5.13.0

